Update number from NumberEntry from int float: - update test to test float - update NumberEntryFormModel to float - update input field to 0.01-step input

+++ src/main/java/de/thb/guessingaverage/controller/form/NumberEntryFormModel.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class NumberEntryFormModel {
-    private int number;
+    private float number;
 }
